+++ sushii-2/src/commands/users/avatar.rs
@@ -26,7 +26,15 @@ async fn avatar(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
         }
     };
 
-    let user = target_id.to_user(ctx).await?;
+    let user = match target_id.to_user(&ctx).await {
+        Ok(u) => u,
+        Err(_) => {
+            msg.reply(&ctx, "Error: Failed to fetch user, are you using a correct user ID?")
+                .await?;
+
+            return Ok(());
+        }
+    };
 
     msg.channel_id
         .send_message(ctx, |m| {