+++ sushii-2/src/commands/users/rep.rs
@@ -36,7 +36,15 @@ async fn rep(ctx: &Context, msg: &Message, mut args: Args) -> CommandResult {
         return Ok(());
     }
 
-    let target_user = target_id.to_user(&ctx).await?;
+    let target_user = match target_id.to_user(&ctx).await {
+        Ok(u) => u,
+        Err(_) => {
+            msg.reply(&ctx, "Error: Failed to fetch user, are you using a correct user ID?")
+                .await?;
+
+            return Ok(());
+        }
+    };
 
     if target_user.bot {
         msg.channel_id