+++ sushii-2/src/commands/users/userinfo.rs
@@ -37,7 +37,16 @@ async fn userinfo(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
         }
     };
 
-    let user = target_id.to_user(ctx).await?;
+    let user = match target_id.to_user(&ctx).await {
+        Ok(u) => u,
+        Err(_) => {
+            msg.reply(&ctx, "Error: Failed to fetch user, are you using a correct user ID?")
+                .await?;
+
+            return Ok(());
+        }
+    };
+
     let member = guild_id.member(ctx, target_id).await;
     let now = Utc::now();